Changed: Moved the examples to tests folder.
