+++ src/Examples/Jobs/README.md
@@ -15,7 +15,7 @@ you want to handle the execution logic in the same method where you call
 The job definition serves as context/data, while the actual work is done in the
 lambda callback right where you need it.
 
-See [**SimpleJobDefinitionExample.cs**](./SimpleJobDefinitionExample.cs).
+See [**SimpleJobDefinitionExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/SimpleJobDefinitionExample.cs).
 
 ### Self-Executing Job
 
@@ -25,7 +25,7 @@ forget.
 The fields you store in the job definition are effectively method parameters
 passed in to run the job.
 
-See [**SelfExecutingJobExample.cs**](./SelfExecutingJobExample.cs).
+See [**SelfExecutingJobExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/SelfExecutingJobExample.cs).
 
 ## Best Practices
 
@@ -47,16 +47,16 @@ natural breakpoints.
 UI dialogs) that can cancel. Propagate external `OperationCanceledException`
 through `context.CancelAndThrow()`.
 
-See [**CancellationExample.cs**](./BestPractices/CancellationExample.cs) and
-[**InteractiveCancellationExample.cs**](./BestPractices/InteractiveCancellationExample.cs).
+See [**CancellationExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/BestPractices/CancellationExample.cs) and
+[**InteractiveCancellationExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/BestPractices/InteractiveCancellationExample.cs).
 
 ### Progress Reporting
 
 Use percentage for determinate progress when total work is known. For
 indeterminate progress, use `Size.One` as maximum to avoid division by zero.
 
-See [**DeterminateProgressExample.cs**](./BestPractices/DeterminateProgressExample.cs)
-and [**IndeterminateProgressExample.cs**](./BestPractices/IndeterminateProgressExample.cs).
+See [**DeterminateProgressExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/BestPractices/DeterminateProgressExample.cs)
+and [**IndeterminateProgressExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/BestPractices/IndeterminateProgressExample.cs).
 
 ### Factory Methods
 
@@ -67,7 +67,7 @@ starting logic as one.
 Factory methods encapsulate dependency injection and job configuration,
 creating reusable job creation patterns.
 
-See [**FactoryMethodExample.cs**](./BestPractices/FactoryMethodExample.cs).
+See [**FactoryMethodExample.cs**](../../../tests/NexusMods.Jobs.Tests/Examples/BestPractices/FactoryMethodExample.cs).
 
 ## General Guidelines
