+++ src/Examples/Program.cs
@@ -6,6 +6,9 @@ namespace Examples;
 [UsedImplicitly] // Designer stuff
 public class Program
 {
+    [STAThread]
+    public static int Main() => 0;
+    
     /// <summary>
     /// Don't Delete this method. It's used by the Avalonia Designer.
     /// </summary>
