+++ tests/NexusMods.Jobs.Tests/Examples/SelfExecutingJobExample.cs
@@ -4,7 +4,7 @@ using NexusMods.Abstractions.Jobs;
 using NexusMods.Paths;
 using Xunit;
 // ReSharper disable UnusedParameter.Local
-namespace Examples.Jobs;
+namespace NexusMods.Jobs.Tests.Examples;
 
 // When running more complex jobs, you will want to make dedicated job objects.
 // For self-contained jobs, use `IJobDefinitionWithStart`.
