+++ tests/NexusMods.Jobs.Tests/Startup.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using NexusMods.Paths;
 
 namespace NexusMods.Jobs.Tests;
 
@@ -7,8 +8,15 @@ public class Startup
 {
     public void ConfigureServices(IServiceCollection container)
     {
-        container.AddJobMonitor()
-            .AddLogging(builder => builder.AddXUnit());
+        const KnownPath baseKnownPath = KnownPath.EntryDirectory;
+        var baseDirectory = $"NexusMods.Examples.Tests-{Guid.NewGuid()}";
+        var prefix = FileSystem.Shared.GetKnownPath(baseKnownPath).Combine(baseDirectory);
+
+        container
+            .AddFileSystem()
+            .AddSingleton<TemporaryFileManager>(_ => new TemporaryFileManager(FileSystem.Shared, prefix))
+            .AddJobMonitor()
+            .AddSingleton<HttpClient>();
     }
 }
